Update normalizationSection.wiki
this should do it. When there is a need it can even be shortened. I think for a best pratices paper this should be the level of detail. And that mainly the favourite methods should be mentioned. Details about the alternative methods and in-depth comparison can be found easily by reader by looking at the citedt reviews? Do you aggree?